KrushiAI-Disease-Recognition/requirements.txt
streamlit==1.49.1
numpy==2.3.3
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
scikit-learn==1.7.2
tensorflow==2.20.0
Pillow==11.3.0
plotly==6.3.0
streamlit-option-menu==0.4.0
opencv-python-headless==4.11.0.86
requests==2.32.5
altair==5.5.0
toml==0.10.2
